Project3(ECApi):ExceptionHandler Middleware yazıldı.

+++ ECommerceAPI/Core/ECommerceAPI.Application/Features/CartsItems/Commands/Delete/DeleteCartItemCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using ECommerceAPI.Application.UnitOfWork;
+using ECommerceAPI.Domain.Exceptions;
 using FluentValidation;
 using MediatR;
 using System;
@@ -23,6 +24,12 @@ namespace ECommerceAPI.Application.Features.CartsItems.Commands.Delete
         public async Task<DeleteCartItemCommandResponse> Handle(DeleteCartItemCommand request, CancellationToken cancellationToken)
         {
             await _validator.ValidateAsync(request, cancellationToken);
+
+            var cartItem = await _unitOfWork.CartItemRepository.GetByIdAsync(request.Id);
+            if (cartItem == null)
+            {
+                throw new NotFoundException($"{request.Id} Id'sine ait Sepet İçeriği bulunamadı...");
+            }
 
             await _unitOfWork.CartItemRepository.RemoveAsync(request.Id);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
